tabu search working finaly

+++ src/DataModel/Domain/AmbulanceType.cs
@@ -3,6 +3,6 @@ namespace ESSP.DataModel;
 public readonly struct AmbulanceType
 {
   public string Name { get; init; }
-  public int Cost { get; init; }
+  public double Cost { get; init; }
 }
